+++ src/math/mod.rs
@@ -5,6 +5,7 @@
 
 #[macro_use]
 mod macros;
+mod cast;
 mod displacement;
 mod normal;
 mod orientation;
@@ -12,6 +13,10 @@ mod position;
 mod screen_position;
 mod velocity;
 
+pub use cast::ToF32;
+pub use cast::ToI32;
+pub use cast::ToU32;
+pub use cast::ToUsize;
 pub use displacement::Displacement;
 pub use normal::Normal;
 pub use normal::ZeroLengthError;