+++ src/math/position.rs
@@ -25,6 +25,7 @@ semantic_newtype!(
 );
 
 #[cfg(test)]
+#[allow(clippy::float_cmp)]
 mod tests {
     use super::*;